feat: implement theme toggle functionality with custom hook
